requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
Django==2.2.17
django-celery==3.3.1
django-celery-beat==2.1.0
django-filter==2.4.0
django-timezone-field==4.0
djangorestframework==3.12.2
importlib-metadata==2.1.0
kombu==3.0.37
migrate==0.3.8
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
vine==1.3.0
zipp==1.2.0
